Big rename from Atom to Enum.
Elements like `nil, `foo, `true, … that used to be called Atoms are now called
Enums. This prevent an unfortunate clash with the concept of atoms of a DNF
which is what is stored in the nodes of BDDs. Variant (like polymorphic variants)
were also considered, but they lead to ambiguous namings when used together with
polymorphic variables (like VarVariant Bdds ?).