Tweak the priorities a bit.
With lower priorities for Regexp operators (;, *, +, ?) the pretty printer will
choose to not print parentheses when encountering:

  (x & y)*

since the priorities of * are lower than those of &
and then x & y * is not valid syntax, since it mixes a type (x) and a regexp (y*).
We rise the priorities of those operators above negation. The priority for | (in regexp)
remains the same as the one of | (in types) since even if we choose to print one or the
other, it will be the same symbol, and reparsed correctly.