Fix pretty-printing of regexp.
Internaly, for types of the form

(t1, X1) | (t2, X2) | ... (tn, Xn)
where the Xi are sequences types, an automaton is generated with transitions
q -> ti -> qXi.

This union of product is generated using Cartesian.normal, which makes products
disjoint on their first component. This is a problem for variables, since
the union:
('a, X1) | ('b, X2) is now rendered as:

('a\'b, ...) | ('b\'a, ...) | ('a&'b, ...) since an intersection of variable is
never empty. This then goes on recursively for the second component, yielding a
 blow-up in size of the union. In a way, this is generating a deterministic automaton.

This is unneeded. We use Cartesian.simple which just removes the negative part of the
products, if any. In practice this may yield a non-det automaton, which is perfectly
ok for a type.